- fixing issue with 2.5 schema (4.8) not being updated to 3.0 (was 4.8, now 4.9) - fixing issue with migrating in-memory indexes
