+++ Raven.Database/Raven.Database.csproj
@@ -957,7 +957,8 @@
     <Compile Include="Storage\Esent\SchemaUpdates\Updates\From44To45.cs" />
     <Compile Include="Storage\Esent\SchemaUpdates\Updates\From45To46.cs" />
     <Compile Include="Storage\Esent\SchemaUpdates\Updates\From46To47.cs" />
-    <Compile Include="Storage\Esent\SchemaUpdates\Updates\From47To48.cs">
+    <Compile Include="Storage\Esent\SchemaUpdates\Updates\From47To48.cs" />
+    <Compile Include="Storage\Esent\SchemaUpdates\Updates\From48To49.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="Storage\Esent\StorageActionsAccessor.cs" />
