+++ Raven.Database/Storage/Esent/SchemaCreator.cs
@@ -13,7 +13,7 @@ namespace Raven.Storage.Esent
 	[CLSCompliant(false)]
 	public class SchemaCreator
 	{
-	    public const string SchemaVersion = "4.8";
+	    public const string SchemaVersion = "4.9";
 		private readonly Session session;
 
 		public SchemaCreator(Session session)
